Family Program
In this program it'll ask you questions like 'What is you're dad called' or 'What is you're mom called'. This is a very simple programs, and i wrote it because i am a beginner and i will be happy if you will feedback me on my program. Then i can be better at programming in the future.